requirements.txt
streamlit
pandas
camelot-py[cv]
paddleocr
paddlepaddle
pdf2image
opencv-python
pillow
tqdm
ghostscript
